safe lstr* replacements: mir_strlen, mir_wstrlen, mir_strcpy, mir_wstrcpy, mir_strncpy, mir_wstrncpy, mir_strcat, mir_wstrcat, mir_strncat, mir_wstrncat
git-svn-id: http://svn.miranda-ng.org/main/trunk@11171 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Clist_modern/src/modern_clcopts.cpp
@@ -1574,7 +1574,7 @@ static INT_PTR CALLBACK DlgProcClcBkgOpts(HWND hwndDlg, UINT msg, WPARAM wParam,
 static INT_PTR BkgrCfg_Register(WPARAM wParam, LPARAM lParam)
 {
 	char *szSetting = (char*)wParam;
-	int len = strlen(szSetting) + 1;
+	size_t len = strlen(szSetting) + 1;
 
 	char *value = (char *)mir_alloc(len + 4); // add room for flags (DWORD)
 	memcpy(value, szSetting, len);
